removed comments and general cleanup in the codebase
still some fine tuning to do, but I consider it ...oh, damn, I know there are some issues, but I've been working on this for weeks haha I need to show or I'm gonna feel procrastinating on the delivery. I guess this is enough to gauge my abilities against what you're looking for. I hope you like it, it was a labor of love. :-)